+++ dart_sample/lib/sample_routes.dart
@@ -42,6 +42,7 @@ class SampleRoutes {
   static const String circularProgressIndicator =
       '/circular-progress-indicator';
   static const String refreshIndicator = '/refresh-indicator';
+  static const String barControls = '/bar-controls';
   static const String slider = '/slider';
   static const String rangeSlider = '/range-slider';
   static const String card = '/card';