+++ src/Plumix/Rendering/Object.PaintingContext.cs
@@ -151,6 +151,18 @@ public sealed class PaintingContext
         });
     }
 
+    public void DrawGeometry(IBrush? brush, IPen? pen, Geometry geometry)
+    {
+        ArgumentNullException.ThrowIfNull(geometry);
+        var pictureLayer = EnsurePictureLayer();
+        pictureLayer.AddDrawCommand((drawingContext, sceneOffset) =>
+        {
+            using var transform = drawingContext.PushTransform(
+                Matrix.CreateTranslation(sceneOffset.X, sceneOffset.Y));
+            drawingContext.DrawGeometry(brush, pen, geometry);
+        });
+    }
+
     public void DrawTextLayout(TextLayout layout, Point point)
     {
         var pictureLayer = EnsurePictureLayer();
@@ -270,6 +282,22 @@ public sealed class PaintingContext
             BorderRadius = borderRadius
         };
 
+        _containerLayer.Append(layer);
+
+        var childContext = new PaintingContext(layer);
+        painter(childContext);
+        childContext.StopRecordingIfNeeded();
+    }
+
+    public void PushClipGeometry(Geometry geometry, Action<PaintingContext> painter)
+    {
+        ArgumentNullException.ThrowIfNull(geometry);
+        StopRecordingIfNeeded();
+
+        var layer = new ClipGeometryLayer
+        {
+            Geometry = geometry
+        };
         _containerLayer.Append(layer);
 
         var childContext = new PaintingContext(layer);